better structure for routing
